Changed the player postition from int to double appened some code back into int

+++ src/com/redomar/game/entities/Dummy.java
@@ -31,13 +31,13 @@ public class Dummy extends Mob {
 
 		List<Player> players = level.getPlayers(this, 10);
 		if (players.size() > 0) {
-			followMovementAI(getX(), getY(), Game.getPlayer().getX(), Game
+			followMovementAI((int) getX(), (int) getY(), (int) Game.getPlayer().getX(), (int) Game
 					.getPlayer().getY(), xa, ya, this);
 		}else{
 			isMoving = false;
 		}
 
-		setSwim(new Swim(level, getX(), getY()));
+		setSwim(new Swim(level, (int) getX(), (int) getY()));
 		swimType = getSwim().swimming(isSwimming, isMagma, isMuddy);
 		isSwimming = swimType[0];
 		isMagma = swimType[1];
@@ -70,8 +70,8 @@ public class Dummy extends Mob {
 		}
 
 		int modifier = 8 * scale;
-		int xOffset = getX() - modifier / 2;
-		int yOffset = getY() - modifier / 2 - 4;
+		int xOffset = (int) getX() - modifier / 2;
+		int yOffset = (int) getY() - modifier / 2 - 4;
 
 		if (isSwimming || isMagma || isMuddy) {
 			swimColour = getSwim().waveCols(isSwimming, isMagma, isMuddy);
